feat: Block Component를 Input Component를 사용하도록 수정(#23)
